requirements.txt
# Core dependencies
pydantic>=2.0.0
openai>=1.0.0
PyYAML>=6.0
mutagen>=1.47.0

# Optional dependencies for enhanced functionality
# musicbrainzngs>=0.7.1  # For MusicBrainz integration
# discogs-client>=2.3.0  # For Discogs integration